requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.3
python-decouple==3.8
dj-database-url==2.1.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
djangorestframework-simplejwt==5.3.1
Pillow>=10.0,<11.0
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
redis==5.0.1
cryptography==42.0.0